pyproject.toml
[project]
name = "chemml-purple-agent"
version = "1.0.0"
description = "A competition-grade ML engineering Purple Agent for AgentBeats MLE-Bench — with analytical chemistry specialization"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk[http-server]>=0.3.20",
    "uvicorn>=0.38.0",
    "httpx>=0.28.1",
    "litellm>=1.0.0",
    "python-dotenv>=1.0.0",
    # Data science
    "pandas>=2.2.0",
    "numpy>=2.0.0",
    "scikit-learn>=1.6.0",
    "xgboost>=2.1.0",
    "lightgbm>=4.5.0",
    "scipy>=1.14.0",
    # Deep learning
    "torch>=2.5.0",
    "torchvision>=0.20.0",
    "Pillow>=11.0.0",
    # Chemistry (optional but gives competitive edge)
    "rdkit>=2024.3.1",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.1",
]
